finished changes to run usual DDPM conditioned on one hot encoded segmentation map

+++ tta_uia_segmentation/src/models/ConditionalGaussianDiffusion.py
@@ -70,7 +70,7 @@ class ConditionalGaussianDiffusion(GaussianDiffusion):
         max_t = default(max_t, self.num_timesteps)
         t = torch.randint(min_t, max_t, (b,), device=device).long()
         img = self.normalize(img)
-        cond_img = self.normalize(cond_img)  
+        cond_img = self.normalize(cond_img.type(torch.int8))  
         
         return self.p_losses_conditioned_on_img(img, t, cond_img, *args, **kwargs)
